requirements.txt
django
djangorestframework
markdown
django-filter
elasticsearch
elasticsearch-dsl
django-elasticsearch-dsl
python-dotenv